+++ lib/src/components/alert.dart
@@ -3,6 +3,7 @@ import 'package:shadcn_ui/src/components/image.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
 import 'package:shadcn_ui/src/utils/border.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 
 enum ShadAlertVariant {
   primary,
@@ -25,6 +26,7 @@ class ShadAlert extends StatelessWidget {
     this.descriptionStyle,
     this.mainAxisAlignment,
     this.crossAxisAlignment,
+    this.orderPolicy,
   }) : variant = ShadAlertVariant.primary;
 
   const ShadAlert.destructive({
@@ -42,6 +44,7 @@ class ShadAlert extends StatelessWidget {
     this.descriptionStyle,
     this.mainAxisAlignment,
     this.crossAxisAlignment,
+    this.orderPolicy,
   }) : variant = ShadAlertVariant.destructive;
 
   const ShadAlert.raw({
@@ -60,6 +63,7 @@ class ShadAlert extends StatelessWidget {
     this.descriptionStyle,
     this.mainAxisAlignment,
     this.crossAxisAlignment,
+    this.orderPolicy,
   });
 
   final ShadAlertVariant variant;
@@ -76,6 +80,12 @@ class ShadAlert extends StatelessWidget {
   final TextStyle? descriptionStyle;
   final MainAxisAlignment? mainAxisAlignment;
   final CrossAxisAlignment? crossAxisAlignment;
+
+  /// {@template ShadAlert.orderPolicy}
+  /// The order policy of the items that compose the alert, defaults to
+  /// [WidgetOrderPolicy.linear()].
+  /// {@endtemplate}
+  final WidgetOrderPolicy? orderPolicy;
 
   @override
   Widget build(BuildContext context) {
@@ -142,6 +152,10 @@ class ShadAlert extends StatelessWidget {
         effectiveAlertTheme.crossAxisAlignment ??
         CrossAxisAlignment.start;
 
+    final effectiveOrderPolicy = orderPolicy ??
+        effectiveAlertTheme.orderPolicy ??
+        const WidgetOrderPolicy.linear();
+
     return ShadDecorator(
       decoration: effectiveDecoration,
       child: Row(
@@ -150,7 +164,7 @@ class ShadAlert extends StatelessWidget {
         textDirection: textDirection,
         children: [
           if (effectiveIcon != null) effectiveIcon,
-          Flexible(
+          Expanded(
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               mainAxisSize: MainAxisSize.min,
@@ -168,7 +182,7 @@ class ShadAlert extends StatelessWidget {
               ],
             ),
           ),
-        ],
+        ].order(effectiveOrderPolicy),
       ),
     );
   }