+++ lib/src/components/form/fields/switch.dart
@@ -3,6 +3,7 @@ import 'package:shadcn_ui/src/components/form/field.dart';
 import 'package:shadcn_ui/src/components/switch.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 
 class ShadSwitchFormField extends ShadFormBuilderField<bool> {
   ShadSwitchFormField({
@@ -32,6 +33,9 @@ class ShadSwitchFormField extends ShadFormBuilderField<bool> {
     Widget? inputSublabel,
     EdgeInsets? padding,
     TextDirection? direction,
+
+    /// {@macro ShadSwitch.orderPolicy}
+    WidgetOrderPolicy? orderPolicy,
   }) : super(
           initialValue: initialValue,
           onChanged: (!enabled || onChanged == null)
@@ -65,6 +69,7 @@ class ShadSwitchFormField extends ShadFormBuilderField<bool> {
               thumbColor: thumbColor,
               uncheckedTrackColor: uncheckedTrackColor,
               checkedTrackColor: checkedTrackColor,
+              orderPolicy: orderPolicy,
             );
           },
         );