+++ lib/src/theme/components/radio.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 import 'package:flutter/widgets.dart';
 
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 
 @immutable
 class ShadRadioTheme {
@@ -20,6 +21,7 @@ class ShadRadioTheme {
     this.alignment,
     this.runAlignment,
     this.crossAxisAlignment,
+    this.orderPolicy,
   });
 
   final bool merge;
@@ -53,6 +55,9 @@ class ShadRadioTheme {
 
   /// {@macro ShadRadioGroup.crossAxisAlignment}
   final WrapCrossAlignment? crossAxisAlignment;
+
+  /// {@macro ShadRadio.orderPolicy}
+  final WidgetOrderPolicy? orderPolicy;
 
   static ShadRadioTheme lerp(
     ShadRadioTheme a,
@@ -91,6 +96,7 @@ class ShadRadioTheme {
     WrapAlignment? alignment,
     WrapAlignment? runAlignment,
     WrapCrossAlignment? crossAxisAlignment,
+    WidgetOrderPolicy? orderPolicy,
   }) {
     return ShadRadioTheme(
       merge: merge ?? this.merge,
@@ -106,6 +112,7 @@ class ShadRadioTheme {
       alignment: alignment ?? this.alignment,
       runAlignment: runAlignment ?? this.runAlignment,
       crossAxisAlignment: crossAxisAlignment ?? this.crossAxisAlignment,
+      orderPolicy: orderPolicy ?? this.orderPolicy,
     );
   }
 
@@ -125,6 +132,7 @@ class ShadRadioTheme {
       alignment: other.alignment,
       runAlignment: other.runAlignment,
       crossAxisAlignment: other.crossAxisAlignment,
+      orderPolicy: other.orderPolicy,
     );
   }
 
@@ -145,7 +153,8 @@ class ShadRadioTheme {
         other.runSpacing == runSpacing &&
         other.alignment == alignment &&
         other.runAlignment == runAlignment &&
-        other.crossAxisAlignment == crossAxisAlignment;
+        other.crossAxisAlignment == crossAxisAlignment &&
+        other.orderPolicy == orderPolicy;
   }
 
   @override
@@ -162,6 +171,7 @@ class ShadRadioTheme {
         runSpacing.hashCode ^
         alignment.hashCode ^
         runAlignment.hashCode ^
-        crossAxisAlignment.hashCode;
+        crossAxisAlignment.hashCode ^
+        orderPolicy.hashCode;
   }
 }